Rename the types module in mytardis_client to objects, to prepare for a later change
