+++ src/inspector/inspector.py
@@ -9,7 +9,7 @@ from src.blueprints.datafile import RawDatafile
 from src.config.config import ConfigFromEnv
 from src.crucible.crucible import Crucible
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
-from src.mytardis_client.types import MyTardisObject
+from src.mytardis_client.objects import MyTardisObject
 from src.overseers.overseer import Overseer
 from src.smelters.smelter import Smelter
