+++ src/overseers/overseer.py
@@ -14,7 +14,7 @@ from src.blueprints.custom_data_types import URI
 from src.config.config import IntrospectionConfig
 from src.mytardis_client.endpoints import get_endpoint
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
-from src.mytardis_client.types import MyTardisObject, get_type_info
+from src.mytardis_client.objects import MyTardisObject, get_type_info
 from src.utils.types.singleton import Singleton
 
 logger = logging.getLogger(__name__)
