+++ tests/ignored/ignoredtest_ingestion_factory.py
@@ -38,7 +38,7 @@ from src.forges.forge import Forge
 from src.ingestion_factory import IngestionFactory
 from src.ingestion_factory.factory import IngestionResult
 from src.mytardis_client.mt_rest import MyTardisRESTFactory
-from src.mytardis_client.types import MyTardisObject
+from src.mytardis_client.objects import MyTardisObject
 from src.overseers.overseer import Overseer
 from src.smelters.smelter import Smelter
